Use a view plugin, not tooltips
