+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from "./inline-edit.js";
 export * from "./state.js";
 export * from "./theme.js";
+export * from "./trigger.js";
